Task4.py

+++ Task4.py
@@ -1,3 +1,3 @@
 #4th program
 Number = '123.456'
-print((float(Number) * 10) % 10)
+print(int(float(Number) * 10) % 10)
